+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## master
+## 0.0.15
 - Handle switch state change for PushOnButton
 - Added Domoticz devices to represent Xiaomi Cube rotation angle and current side
 - Added Domoticz devices to represent temperature, humidity and pressure in one device
+- Added auth support for connection to MQTT server
 
 New devices:
 - Commercial Electric Matte White Recessed Retrofit Smart Led Downlight - 4 Inch